move bin/api_usage.dart impl into lib/
