+++ packages/corpus/bin/api_usage.dart
@@ -5,19 +5,10 @@
 import 'dart:io';
 
 import 'package:args/args.dart';
-import 'package:pub_semver/pub_semver.dart';
-import 'package:cli_util/cli_logging.dart';
-import 'package:corpus/api.dart';
-import 'package:corpus/cache.dart';
-import 'package:corpus/packages.dart';
-import 'package:corpus/pub.dart';
-import 'package:corpus/report.dart';
-import 'package:corpus/surveyor.dart';
-import 'package:path/path.dart' as path;
-import 'package:pub_semver/pub_semver.dart';
+import 'package:corpus/api_usage.dart';
 
 void main(List<String> args) async {
-  var argParser = createArgParser();
+  var argParser = _createArgParser();
 
   late ArgResults argResults;
   try {
@@ -36,90 +27,17 @@ void main(List<String> args) async {
 
   final packageName = argResults.rest.first;
   final packageLimit =
-      int.tryParse(argResults['package-limit'] ?? '') ?? 0x7fffffff;
+      int.tryParse(argResults['package-limit'] ?? '') ?? 0x7FFFFFFF;
   bool showSrcReferences = argResults['show-src-references'] as bool;
 
-  var log = Logger.standard();
-
-  log.stdout('API usage analysis for package:$packageName.');
-  log.stdout('');
-
-  var pub = Pub();
-  var packageManager = PackageManager();
-
-  var progress = log.progress('querying pub.dev');
-
-  var targetPackage = await pub.getPackageInfo(packageName);
-
-  var packageStream = pub.popularDependenciesOf(packageName);
-
-  progress.finish(showTiming: true);
-
-  List<ApiUsage> usageInfo = [];
-
-  var count = 0;
-
-  await for (var package in packageStream) {
-    log.stdout('');
-    log.stdout('${package.name} v${package.version}');
-
-    // Skip a package when its constraints don't include the latest stable.
-    var constraint = package.constraintFor(targetPackage.name);
-    if (constraint == null) {
-      log.stdout('skipping - no constraint on ${targetPackage.name}');
-      continue;
-    }
-    if (!constraint.allows(Version.parse(targetPackage.version))) {
-      log.stdout(
-          "skipping - version dep ($constraint) doesn't support the current "
-          'stable (${targetPackage.version})');
-      continue;
-    }
-
-    bool downloadSuccess =
-        await packageManager.retrievePackageArchive(package, logger: log);
-    if (!downloadSuccess) {
-      log.stdout('error downloading ${package.archiveUrl}');
-      continue;
-    }
-
-    var localPackage = await packageManager.rehydratePackage(package);
-
-    var pubSuccess =
-        await localPackage.pubGet(checkUpToDate: true, logger: log);
-    if (!pubSuccess) {
-      continue;
-    }
-
-    count++;
-
-    progress = log.progress('analyzing package');
-    var usage =
-        await analyzePackage(targetPackage, package, localPackage.directory);
-    progress.finish(message: usage.describeUsage());
-
-    usageInfo.add(usage);
-
-    if (count >= packageLimit) {
-      break;
-    }
-  }
-
-  var report = Report(targetPackage);
-  var file = report.generateReport(
-    usageInfo,
+  await analyzeUsage(
+    packageName: packageName,
+    packageLimit: packageLimit,
     showSrcReferences: showSrcReferences,
   );
-
-  log.stdout('');
-  log.stdout('wrote ${file.path}.');
-
-  packageManager.close();
-
-  pub.close();
 }
 
-ArgParser createArgParser() {
+ArgParser _createArgParser() {
   var parser = ArgParser();
   parser.addFlag(
     'help',
@@ -146,38 +64,4 @@ void printUsage(ArgParser argParser) {
   print('');
   print('options:');
   print(argParser.usage);
-}
-
-Future<ApiUsage> analyzePackage(
-  PackageInfo targetPackage,
-  PackageInfo usingPackage,
-  Directory usingPackageDir,
-) async {
-  var cache = Cache();
-  var file = File(path.join(
-    cache.usageDir.path,
-    '${targetPackage.name}-${targetPackage.version}',
-    '${usingPackage.name}-${usingPackage.version}.json',
-  ));
-
-  if (file.existsSync()) {
-    ApiUsage usage = ApiUsage.fromFile(usingPackage, file);
-    return usage;
-  }
-
-  var apiUsageCollector =
-      ApiUseCollector(targetPackage, usingPackage, usingPackageDir);
-
-  var driver = SurveyorDriver.fromDirs(
-    directories: [usingPackageDir],
-    visitor: apiUsageCollector,
-    excludedPaths: ['example'],
-  );
-
-  await driver.analyze();
-
-  var usage = apiUsageCollector.usage;
-  file.parent.createSync();
-  usage.toFile(file);
-  return usage;
 }
